backend/requirements.txt
audioop-lts==0.2.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pydub==0.25.1
Werkzeug==3.1.3
librosa==0.10.2
numpy==1.26.4
scipy==1.13.1
soundfile==0.12.1
supabase-py==2.4.0
python-dotenv==1.0.1